src/message_player/requirements.txt
paho-mqtt==2.1.0
python-dotenv
python-statemachine